ci: move release code to a separate file
